Update to ares v114r25 release.
  - PlayStation: implemented all GTE instructions (though certainly with
    many serious bugs remaining)
  - ares: improved Debug class notifications

+++ ares/ps1/cpu/core/core.hpp
@@ -8,12 +8,12 @@
   auto powerCore(bool reset) -> void;
 
   //memory.cpp
-  auto readByte(u32 address) ->  u8;
-  auto readHalf(u32 address) -> u16;
+  auto readByte(u32 address) -> u32;
+  auto readHalf(u32 address) -> u32;
   auto readWord(u32 address) -> u32;
 
-  auto writeByte(u32 address,  u8 data) -> void;
-  auto writeHalf(u32 address, u16 data) -> void;
+  auto writeByte(u32 address, u32 data) -> void;
+  auto writeHalf(u32 address, u32 data) -> void;
   auto writeWord(u32 address, u32 data) -> void;
 
   //decoder.cpp
