+++ ares/ps1/cpu/core/memory.cpp
@@ -1,9 +1,9 @@
-inline auto CPU::readByte(u32 address) -> u8 {
+inline auto CPU::readByte(u32 address) -> u32 {
   if(scc.status.cache.isolate) return cache.readByte(address);
   return bus.readByte(address);
 }
 
-inline auto CPU::readHalf(u32 address) -> u16 {
+inline auto CPU::readHalf(u32 address) -> u32 {
   if(scc.status.cache.isolate) return cache.readHalf(address);
   return bus.readHalf(address);
 }
@@ -13,12 +13,12 @@ inline auto CPU::readWord(u32 address) -> u32 {
   return bus.readWord(address);
 }
 
-inline auto CPU::writeByte(u32 address, u8 data) -> void {
+inline auto CPU::writeByte(u32 address, u32 data) -> void {
   if(scc.status.cache.isolate) return cache.writeByte(address, data);
   return bus.writeByte(address, data);
 }
 
-inline auto CPU::writeHalf(u32 address, u16 data) -> void {
+inline auto CPU::writeHalf(u32 address, u32 data) -> void {
   if(scc.status.cache.isolate) return cache.writeHalf(address, data);
   return bus.writeHalf(address, data);
 }
